Tinkered around with the tooling
Added a Verified constructor to the Result type. This is for puzzles
that I have already solved and have gotten a correct answer back from
the advent of code server. This will add a checkmark to the output and
will also ensure I don't break any solutions should I decide to refactor